feat: 혜택 직접 검색 API 구현 (중앙부처)

+++ src/main/java/salpim/umc10thsalpim/domain/benefit/controller/BenefitController.java
@@ -27,8 +27,8 @@ public class BenefitController {
     public ApiResponse<CursorResDTO.Pagination<BenefitResDTO.WelfareSearchResultDTO>> getSearchResult
     (
             @RequestParam String searchKey,
-            @RequestParam List<Integer> regionIds,
-            @RequestParam List<Integer> categoryIds,
+            @RequestParam List<Long> regionIds,
+            @RequestParam List<Long> categoryIds,
             @RequestParam(name="cursor", defaultValue = "-1") String cursor,
             @RequestParam(name="pageSize", defaultValue = "10") Integer pageSize,
             @RequestParam(name="sort", defaultValue = "popular") String sort
